refactor(weather): Flatten repository structure for weather domain
- Moved repository files out of nested repositories/ directory
- Updated import paths to reflect new structure
- Applies to Firestore, Unified, and API-based weather repositories
