+++ lib/data/repositories/weather/weather_repository_impl.dart
@@ -4,7 +4,7 @@ import '../../../../core/logger/app_log.dart';
 import '../../../../core/network/api_error.dart';
 import '../../../../core/network/response/result.dart';
 import '../../../../domain/models/weather/forecast.dart';
-import '../../../services/weather/services/i_weather_api_client.dart';
+import '../../services/weather/i_weather_api_client.dart';
 import 'weather_repository.dart';
 
 /// 天気データを取得するリポジトリの実装クラス。
